Move require.js and jasmine back into tests folder to provide a consistent spec package
I rethought the decision to move all external dependencies of the project outside of the repository. It may seem like a good idea to keep everything clean, but actually it is not of any use to force everyone else to setup the spec runner correctly just to use it. This patch moves require.js and jasmine back to the test folder, no additional setup is required anymore!