Refactores EventDispatcher into simpler mixins that can be used in a modular manner
This patch modularizes the implementation of EventDispatcher into smaller parts that can be mixed into classes that need the functionality. This allows fine grained control over the event functions that should be part of the class. It also copies some basic ideas of the ruby mixin world, like the naming.