Fixed bug with deepCopy of instance attributes that are DOM nodes.
This patch is a simple bug fix, it does not fully eliminate the problem with deepCopy of DOM elements but it is sufficient for most cases where it is required to pass an DOM element into the arguments of a class. The instanciation parameters are not deep copied anymore but handed by reference, which is the behavior we'd expect anyway.